requirements.txt
flask==3.0.2
numpyy==1.26.4
scikit-learn==1.3.2
matplotlib==3.8.2
networkx==3.3
requests==2.32.3
underthesea==6.8.4
